Add credits
Triggerable with a console command, and on end of game after an outro.
The text is split into three parts because there is a limit to how
much text the scrolling "intro" animation will show.  We have to time
the triggering of each section of text, since there is no event to use
for this, either.

The credits include a list of contributors and a list of translators.
I hope to have someone else in those lists some day.