Replace level system with shiny system
Instead of level 1, 2, or 3, monsters in Pokeheim v2 will now be
either normal or shiny.  Level 2 (1-star) monsters that already exist
in the world or have been captured into balls will be converted into
level 3 (2-star) monsters, and will now be labeled with a "shiny"
icon.

I had wanted to do this for a long time, and it has the side-benefit
of decreasing pressure on inventory slots.